refactor(cli): extract comment commands to bin/commands/comment.js
Move `comments`, `comment`, and `comment-delete` into a dedicated module
along with their three rendering helpers (`parseLocationOptions`,
`formatBodyBlock`, `buildCommentTree`), which are only used here. The
`comment_create` onError hint and the `comment_delete_cancel` analytics
key are preserved.
